dytallix-lean-launch/tools/ai-risk-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
PyNaCl==1.5.0
